Add support for secure gRPC connections
This commit adds support for secure gRPC connections in the HttpGateway struct. It introduces a new field called "Secure" which indicates whether the connection should be secure or not. If the connection is secure, it uses TLS credentials when dialing the gRPC server. Otherwise, it uses insecure credentials.

Additionally, the code has been refactored to use the appropriate credentials based on the value of "Secure". The gRPC connection is now stored in a separate variable based on the type of credentials used. The connection is closed when the context is cancelled.

The HTTP gateway now uses the correct gRPC connection when registering the gRPC handlers. The serverHealth function also uses the correct connection to check the server's health.

Overall, this commit enhances the HttpGateway struct to support secure gRPC connections, improving the security of the communication between the gateway and the gRPC server.