Fix error handling in GrpcServer.Run()
The error handling in the GrpcServer.Run() method has been improved. Previously, if there was an error while listening for TCP connections, the error message was logged but the method continued to execute. Now, if there is an error, the method will return immediately. Additionally, a check has been added to ensure that the necessary variables (listen, listen.Addr(), and server) are not nil before starting the gRPC server. If any of these variables are nil, an error message will be logged and the method will return. This ensures that the server is properly set up before attempting to serve requests.