Add request ID to middleware log
This commit adds a request ID to the middleware log for each incoming request. The request ID is generated using the Google UUID package and is set as a header in the request. The log message now includes the request ID along with other information such as the method, URI, protocol, host, and headers. This will make it easier to track and debug requests in the logs. Additionally, the log message for the elapsed time now includes the request ID.