Add github.com/google/uuid v1.3.0 dependency
This commit adds the dependency for github.com/google/uuid v1.3.0 to the go.mod file. The dependency is required for the project and will be used for generating universally unique identifiers (UUIDs) in the codebase. The go.sum file has also been updated to include the hash for the v1.3.0 version of the dependency.